Revert "React Proof of Concept (#1482)"
This reverts commit 1a85e8a6fae459bdbb03e62493c0dd074f6c59a8.
