+++ galaxyui/package.json
@@ -21,20 +21,14 @@
     "@angular/platform-browser": "^6.1.1",
     "@angular/platform-browser-dynamic": "^6.1.1",
     "@angular/router": "^6.1.1",
-    "@types/react": "^16.7.18",
-    "@types/react-dom": "^16.0.11",
     "core-js": "^2.5.7",
     "patternfly": "^3.54.1",
     "patternfly-ng": "4.2.0",
-    "patternfly-react": "^2.29.0",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
     "rxjs": "^6.2.2",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^7.2.0",
-    "@angular-devkit/build-angular": "^0.12.1",
+    "@angular-devkit/build-angular": "~0.7.2",
     "@angular/cli": "6.1.2",
     "@angular/compiler-cli": "^6.1.1",
     "@angular/language-service": "^6.1.1",
@@ -46,10 +40,10 @@
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.5",
     "karma-chrome-launcher": "~2.2.0",
+    "prettier": "1.14.0",
     "karma-coverage-istanbul-reporter": "^2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^1.2.0",
-    "prettier": "^1.15.3",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
