+++ galaxyui/src/app/authors/authors.component.html
@@ -1,0 +1,69 @@
+<app-page-header [headerTitle]="pageTitle" [headerIcon]="pageIcon"></app-page-header>
+
+<div id="authors-page">
+    <div class="padding-15">
+        <div class="row">
+            <div class="col-sm-12">
+                <pfng-toolbar
+                    [config]="toolbarConfig"
+                    (onFilterChange)="filterChanged($event)"
+                    (onSortChange)="sortChanged($event)">
+                </pfng-toolbar>
+            </div>
+        </div>
+    </div>
+
+      <div class="row">
+        <div class="col-sm-12">
+            <div id="authors-list">
+                <pfng-list
+                    [config]="listConfig"
+                    [items]="items"
+                    [itemTemplate]="itemTemplate"
+                    (onClick)="handleListClick($event)">
+                    <ng-template #itemTemplate let-item="item">
+                        <div class="list-pf-left">
+                            <div class="img-container">
+                                <table class="img-wrapper">
+                                    <tr>
+                                        <td><img src="{{ item.avatar_url }}"></td>
+                                    </tr>
+                                </table>
+                            </div>
+                        </div>
+                        <div class="list-pf-content-wrapper">
+                            <div class="list-pf-main-content">
+                                <div class="list-pf-title">
+                                    <a [routerLink]="['/', item.name]" class="title-link"
+                                        tooltip="View author profile and content">{{item.name}}</a>
+                                </div>
+                                <div class="list-pf-description text-overflow-pf">{{item.description }}</div>
+                            </div>
+                            <div class="list-pf-additional-content">
+                                <div *ngFor="let count of item.contentCounts" class="content-count">
+                                     <span class="{{ count.iconClass }}"></span>
+                                    <strong>{{count.count}}</strong> {{count.title}}
+                                </div>
+                            </div>
+                        </div>
+                    </ng-template>
+                </pfng-list>
+            </div>
+        </div>
+    </div>
+
+    <div class="row repository-pagination">
+        <div class="col-sm-12">
+            <div class="pagination" *ngIf="items && items.length && paginationConfig.totalItems > paginationConfig.pageSize">
+                <pfng-pagination
+                    [config]="paginationConfig"
+                    (onPageSizeChange)="handlePageSizeChange($event)"
+                    (onPageNumberChange)="handlePageNumberChange($event)">
+                </pfng-pagination>
+            </div>
+        </div>
+    </div>
+
+</div>
+
+<app-page-loading [loading]="pageLoading"></app-page-loading>
