--- galaxyui/src/app/react/lib/injector-context.ts
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import { Injector } from '@angular/core';
-
-interface IContextDef {
-    injector: Injector;
-}
-
-export const InjectorContext = React.createContext({} as IContextDef);
