--- galaxyui/src/app/react/styles/page-loading.less
@@ -1,23 +1,0 @@
-.page-loading-wrapper {
-    .page-loader {
-        position: fixed;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        z-index: +100;
-        opacity: 1;
-        transition: 1s ease;
-        background: rgba(255, 255, 255, 0.5);
-        .page-loading-inner {
-            position: absolute;
-            top: 50%;
-            left: 50%;
-        }
-    }
-
-    .page-loader.hide-content {
-        opacity: 0;
-        z-index: -100;
-    }
-}
