+++ galaxyui/src/styles.less
@@ -254,16 +254,3 @@ a.failed {
         display: block;
     }
 }
-
-/* React Components */
-// React stylesheets have to be imported here in order to be successfully
-// loaded into JSX
-
-// Page Header
-@import './app/react/styles/page-header.less';
-
-// Community Page
-@import './app/react/styles/community.less';
-
-// Page Loading Widget
-@import './app/react/styles/page-loading.less';
